+++ Plugin.cs
@@ -13,7 +13,8 @@ public class SparrohPlugin : BaseUnityPlugin
 {
     public const string PLUGINGUID = "sparroh.modsettingsmenu";
     public const string PLUGINNAME = "ModSettingsMenu";
-    public const string PLUGINVERSION = "2.1.0";
+    public const string PLUGINVERSION = "2.1.1";
+
 
     public new static ManualLogSource Logger;